+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Desktop - Better display of the QGIS Server version when it's not available
 * Server - Always register the API if possible, raise the error only when requested instead
 * Server - Add the commit ID of QGIS if available in the JSON metadata
+* Server - Set a custom application name when connecting to PostgreSQL on the server side
 
 ## 3.7.4 - 2022-02-15